auth_service/requirements.txt
# Core Flask Dependencies
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1

# Database Drivers
pymysql
SQLAlchemy==2.0.38

# Authentication & Security
PyJWT==2.10.1
cryptography

# Gunicorn for production
gunicorn

# Miscellaneous Utilities
alembic==1.14.1
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
